# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateJobRequest": {
        "properties": {
          "decision_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "holdings_text": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holdings Text"
          },
          "idempotency_key": {
            "maxLength": 100,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "refresh",
              "symbols",
              "quality",
              "backtest",
              "sim_result",
              "sim_compare",
              "decision_plan",
              "filing_check"
            ],
            "title": "Kind",
            "type": "string"
          },
          "portfolio_id": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Id"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Symbols",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "idempotency_key"
        ],
        "title": "CreateJobRequest",
        "type": "object"
      },
...
      "DecisionPolicy": {
        "additionalProperties": false,
        "properties": {
          "constrained_optimizer": {
            "default": false,
            "title": "Constrained Optimizer",
            "type": "boolean"
          },
          "max_invested_pct": {
            "default": 50,
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Max Invested Pct",
            "type": "number"
          },
          "max_position_pct": {
            "default": 5,
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Max Position Pct",
            "type": "number"
          },
          "max_positions": {
            "default": 10,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Max Positions",
            "type": "integer"
          },
          "max_price_age_days": {
            "default": 7,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Max Price Age Days",
            "type": "integer"
          },
          "max_sector_pct": {
            "default": 20,
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Max Sector Pct",
            "type": "number"
          },
          "max_volatility": {
            "default": 0.6,
            "exclusiveMinimum": 0.0,
            "maximum": 10.0,
            "title": "Max Volatility",
            "type": "number"
          },
          "min_coverage": {
            "default": 0.7,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Coverage",
            "type": "number"
          },
          "min_drawdown": {
            "default": -0.5,
            "maximum": 0.0,
            "minimum": -1.0,
            "title": "Min Drawdown",
            "type": "number"
          },
          "min_score": {
            "default": 65,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Min Score",
            "type": "number"
          },
          "trade_threshold_pct": {
            "default": 0.5,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Trade Threshold Pct",
            "type": "number"
          },
          "turnover_penalty": {
            "default": 0,
            "maximum": 0.2,
            "minimum": 0.0,
            "title": "Turnover Penalty",
            "type": "number"
          }
        },
        "title": "DecisionPolicy",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "status": {
            "default": "error",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JobEvent": {
        "properties": {
          "at": {
            "title": "At",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "at",
          "message"
        ],
        "title": "JobEvent",
        "type": "object"
      },
      "JobListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Jobs",
            "type": "array"
          }
        },
        "required": [
          "jobs"
        ],
        "title": "JobListResponse",
        "type": "object"
      },
      "JobResponse": {
        "properties": {
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "checkpoint": {
            "additionalProperties": true,
            "title": "Checkpoint",
            "type": "object"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "events": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/JobEvent"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "result_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Ref"
          },
          "result_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Sha256"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "parameters",
          "origin",
          "status",
          "progress",
          "phase",
          "cancel_requested",
          "created_at",
          "started_at",
          "finished_at",
          "error_code",
          "result_ref",
          "result_sha256",
          "checkpoint"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
      "LocalSettingsResponse": {
        "properties": {
          "keys": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Keys",
            "type": "object"
          },
          "scheduler": {
            "default": "local",
            "title": "Scheduler",
            "type": "string"
          }
        },
        "required": [
          "keys"
        ],
        "title": "LocalSettingsResponse",
        "type": "object"
      },
...
      "ModelResponse": {
        "additionalProperties": false,
        "properties": {
          "blind_validation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blind Validation Id"
          },
          "evidence_note": {
            "default": "Congelar pesos o iniciar seguimiento no acredita una ventaja independiente sobre SPY.",
            "title": "Evidence Note",
            "type": "string"
          },
          "independent_advantage_demonstrated": {
            "const": false,
            "default": false,
            "title": "Independent Advantage Demonstrated",
            "type": "boolean"
          },
          "live_forward_source": {
            "anyOf": [
              {
                "enum": [
                  "blind_validation",
                  "research_lab"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Forward Source"
          },
          "matches_frozen": {
            "title": "Matches Frozen",
            "type": "boolean"
          },
          "mode": {
            "enum": [
              "INVESTOR",
              "RESEARCH"
            ],
            "title": "Mode",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "FROZEN",
              "EXPERIMENTAL",
              "LIVE_FORWARD"
            ],
            "title": "Status",
            "type": "string"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Weights",
            "type": "object"
          },
          "weights_unit": {
            "const": "fraction",
            "default": "fraction",
            "title": "Weights Unit",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "model_id",
          "status",
          "weights",
          "matches_frozen",
          "live_forward_source",
          "blind_validation_id"
        ],
        "title": "ModelResponse",
        "type": "object"
      },
...
      "WeightsRequest": {
        "properties": {
          "momentum": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Momentum",
            "type": "number"
          },
          "quality": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Quality",
            "type": "number"
          },
          "risk": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk",
            "type": "number"
          },
          "value": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "value",
          "quality",
          "momentum",
          "risk"
        ],
        "title": "WeightsRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "GABI local API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/administration/settings": {
      "get": {
        "operationId": "local_settings_api_v1_administration_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Local Settings",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/administration/weights": {
      "post": {
        "operationId": "save_weights_api_v1_administration_weights_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeightsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Save Weights",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "operationId": "create_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Create Job",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-f0-9]{32}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Get Job",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-f0-9]{32}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/result": {
      "get": {
        "operationId": "job_result_api_v1_jobs__job_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-f0-9]{32}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Job Result Api V1 Jobs  Job Id  Result Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Job Result",
        "tags": [
          "administration"
        ]
      }
    },
...
  }
}